generate-factory: FP4/5: Stop flash if first partition fails
If the first bootloader partition fails to flash, end the process there
and warn the user, as it does not bode well for the rest of the process.
Particularly, if this is due to unlock_critical not having been done,
then if we continue to try to flash other partitions, it may result in
a partial flash and brick the device.

device-flasher already handles unlock_critical, so this change primarily
benefits users who try to execute flash scripts directly.

Issue: calyxos#2193
Change-Id: Ie934066bcae38254d3867956bcac17deaf56bde2